+++ sdk/core/System.ClientModel/gen/ModelReaderWriterContextGenerator.cs
@@ -86,7 +86,9 @@ internal sealed partial class ModelReaderWriterContextGenerator : IIncrementalGe
             }
         });
 
-        var cacheProvider = context.CompilationProvider.Select((compilation, _) => new TypeSymbolKindCache());
+        var symbolToKindCacheProvider = context.CompilationProvider.Select((compilation, _) => new TypeSymbolKindCache());
+
+        var symbolToTypeRefCacheProvider = context.CompilationProvider.Select((compilation, _) => new TypeSymbolTypeRefCache());
 
         var attributeInvocations = attributeInfos
             .SelectMany((infos, _) => infos.Where(info => info.Symbol is ITypeSymbol).Select(info => (ITypeSymbol)info.Symbol!))
@@ -107,8 +109,15 @@ internal sealed partial class ModelReaderWriterContextGenerator : IIncrementalGe
             .Combine(modelInfoTypes)
             .Combine(assemblyNameProvider)
             .Combine(referencedTypes)
-            .Combine(cacheProvider)
-            .Select((data, _) => (data.Left.Left.Left.Left, data.Left.Left.Left.Right, data.Left.Left.Right, data.Left.Right, data.Right));
+            .Combine(symbolToKindCacheProvider)
+            .Combine(symbolToTypeRefCacheProvider)
+            .Select((data, _) => (
+                data.Left.Left.Left.Left.Left,
+                data.Left.Left.Left.Left.Right,
+                data.Left.Left.Left.Right,
+                data.Left.Left.Right,
+                data.Left.Right,
+                data.Right));
 
         context.RegisterSourceOutput(combined, ReportDiagnosticAndEmitSource);
     }
@@ -119,7 +128,8 @@ internal sealed partial class ModelReaderWriterContextGenerator : IIncrementalGe
             ImmutableArray<ITypeSymbol> TypeBuilders,
             IAssemblySymbol Assembly,
             ImmutableArray<INamedTypeSymbol> ReferencedContexts,
-            TypeSymbolKindCache SymbolToKindCache) data)
+            TypeSymbolKindCache SymbolToKindCache,
+            TypeSymbolTypeRefCache SymbolToTypeRefCache) data)
     {
         if (data.Contexts.Length > 1)
         {
@@ -139,33 +149,33 @@ internal sealed partial class ModelReaderWriterContextGenerator : IIncrementalGe
         var contextSymbol = data.Contexts[0]!;
 
         string assemblyName = data.Assembly.Name;
-        var contextType = new TypeRef(contextSymbol.Name, contextSymbol.ContainingNamespace.ToDisplayString(), contextSymbol.ContainingAssembly.ToDisplayString());
-
-        var referencedContexts = data.ReferencedContexts.ToImmutableDictionary(
-            refContext => refContext.ContainingAssembly.ToDisplayString(),
-            refContext => new TypeRef(refContext.Name, refContext.ContainingNamespace.ToDisplayString(), refContext.ContainingAssembly.ToDisplayString()));
+        var contextType = data.SymbolToTypeRefCache.Get(contextSymbol, data.SymbolToKindCache);
 
         var builders = data.TypeBuilders
             .SelectMany(typeSymbol => GetRecursiveGenericTypes(typeSymbol, data.SymbolToKindCache))
             .Distinct(SymbolEqualityComparer.Default);
 
-        Dictionary<string, int> dupeCounts = [];
-        Dictionary<ITypeSymbol, string> aliases = [];
-        HashSet<ITypeSymbol> visited = [];
-        var allTypes = data.ReferencedContexts.Concat(data.TypeBuilders);
-        foreach (var typeSymbol in allTypes)
-        {
-            CheckForDupe(typeSymbol, data.SymbolToKindCache, dupeCounts, aliases, visited);
-        }
-        RemoveNonDupe(aliases, dupeCounts);
+        var referencedContexts = data.ReferencedContexts.ToImmutableDictionary(
+            refContext => refContext.ContainingAssembly.ToDisplayString(),
+            refContext => data.SymbolToTypeRefCache.Get(refContext, data.SymbolToKindCache));
 
         var typeGenerationSpecs = builders
             .Select(symbol =>
             {
-                if (symbol is not ITypeSymbol typeSymbol)
+                if (symbol is not ITypeSymbol typeSymbol || typeSymbol.DeclaredAccessibility == Accessibility.Private)
                     return null;
 
-                var type = TypeRef.FromTypeSymbol(typeSymbol, data.SymbolToKindCache, aliases);
+                if (data.SymbolToKindCache.Get(typeSymbol) == TypeBuilderKind.Unknown)
+                    return null;
+
+                var type = data.SymbolToTypeRefCache.Get(typeSymbol, data.SymbolToKindCache);
+                if (type.ObsoleteLevel == ObsoleteLevel.Error)
+                {
+                    // if its marked as error obsolete we can't create a builder for it
+                    // you cannot suppress the reference to an obsolete type marked as error
+                    return null;
+                }
+
                 var itemType = type.GetInnerItemType();
 
                 if (!HasAccessibleParameterlessConstructor(typeSymbol, data.SymbolToKindCache) && itemType.IsSameAssembly(contextType))
@@ -192,7 +202,7 @@ internal sealed partial class ModelReaderWriterContextGenerator : IIncrementalGe
                     Modifier = "internal",
                     Type = type,
                     Kind = data.SymbolToKindCache.Get(typeSymbol),
-                    PersistableModelProxy = proxy is null ? null : TypeRef.FromTypeSymbol(proxy, data.SymbolToKindCache, aliases),
+                    PersistableModelProxy = proxy is null ? null : data.SymbolToTypeRefCache.Get(proxy, data.SymbolToKindCache),
                     ContextType = referencedContexts.ContainsKey(type.Assembly) ? referencedContexts[type.Assembly] : contextType,
                 };
             })
@@ -216,22 +226,6 @@ internal sealed partial class ModelReaderWriterContextGenerator : IIncrementalGe
         OnSourceEmitting?.Invoke(contextGenerationSpec);
         Emitter emitter = new(context);
         emitter.Emit(contextGenerationSpec);
-    }
-
-    private void RemoveNonDupe(Dictionary<ITypeSymbol, string> aliases, Dictionary<string, int> dupeCounts)
-    {
-        List<ITypeSymbol> typesToRemove = [];
-        foreach (var kvp in aliases)
-        {
-            if (dupeCounts.TryGetValue(kvp.Key.Name, out var count) && count == 1)
-            {
-                typesToRemove.Add(kvp.Key);
-            }
-        }
-        foreach (var type in typesToRemove)
-        {
-            aliases.Remove(type);
-        }
     }
 
     private static ImmutableArray<INamedTypeSymbol> GetContextsFromDependencies(Compilation compilation)
@@ -261,32 +255,6 @@ internal sealed partial class ModelReaderWriterContextGenerator : IIncrementalGe
         }
 
         return matchingTypes.ToImmutableArray();
-    }
-
-    private static void CheckForDupe(
-        ITypeSymbol typeSymbol,
-        TypeSymbolKindCache cache,
-        Dictionary<string, int> dupeCounts,
-        Dictionary<ITypeSymbol, string> aliases,
-        HashSet<ITypeSymbol> visited)
-    {
-        if (visited.Contains(typeSymbol))
-            return;
-
-        visited.Add(typeSymbol);
-
-        var itemSymbol = typeSymbol.GetItemSymbol(cache);
-        if (itemSymbol is not null)
-        {
-            CheckForDupe(itemSymbol, cache, dupeCounts, aliases, visited);
-            return;
-        }
-
-        var typeName = typeSymbol.Name;
-        dupeCounts.TryGetValue(typeName, out var count);
-        aliases.Add(typeSymbol, $"{typeName}_{count++}");
-        dupeCounts[typeName] = count;
-        return;
     }
 
     /// <summary>