+++ sdk/core/System.ClientModel/tests/gen.unit/InvocationTests/ListTests.cs
@@ -19,12 +19,6 @@ namespace System.ClientModel.SourceGeneration.Tests.Unit.InvocationTests
             if (invocationDuped)
             {
                 var dupedListModel = ValidateBuilder("TestProject1", expectation, dict);
-                Assert.AreEqual($"List<{expectation.TypeName}_0>", listModel.Type.Alias);
-                Assert.AreEqual($"List<{expectation.TypeName}_1>", dupedListModel.Type.Alias);
-            }
-            else
-            {
-                Assert.IsNull(listModel.Type.Alias);
             }
 
             Assert.IsTrue(dict.TryGetValue($"{expectation.Namespace}.{expectation.TypeName}", out var itemModel));